fixing additional specifier

+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: varaniba <varaniba@student.42.fr>            +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2026/03/10 14:07:55 by varaniba      #+#    #+#                 */
-/*   Updated: 2026/05/04 17:35:30 by varaniba      ########   odam.nl         */
+/*   Updated: 2026/05/04 17:53:34 by varaniba      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -76,6 +76,6 @@ int		ft_putnbr_b(long long nbr, char *base);
 int		ft_putnbr_b_uns(unsigned long long nbr, char *base);
 int		ft_print_float(double nbr, const char *specif);
 void	offset(const char *format_str, int *i);
-int		get_precision(double *nbr, const char *specif, int *multiplier);
+int		get_precision(double *nbr, const char *specif, long *multiplier);
 
 #endif
